migrate individual files in a transaction instead of the entire migra… (#175)
* migrate individual files in a transaction instead of the entire migration

* fix failing test

+++ src/main/java/uk/co/ogauthority/pwa/features/consentdocumentmigration/DocumentMigrationRecordRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentMigrationRecordRepository extends JpaRepository<DocumentMigrationRecord, Long> {
+public interface DocumentMigrationRecordRepository extends JpaRepository<DocumentMigrationRecord, Integer> {
 
   List<DocumentMigrationRecord> findAllByMigrationSuccessfulIsFalse();
